ARIES-1435 Sharing policy updates for dynamic imports, if necessary, should proceed up the subsystem region tree.
Whether or not a sharing policy already allows a matching capability for a dynamic import is not a sufficient condition for terminating the walk up the region tree. It may only be used
for the purposes of avoiding an unnecessary update to a particular sharing policy in the path.

Add test.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1709860 13f79535-47bb-0310-9956-ffa450edef68