[ARIES-1445] Bundles that are not direct dependencies of a subsystem can be removed while still in use
Remove unnecessary code. The special check for a requirement namespace of osgi.service when no wires exist is no longer necessary. Requirements in the osgi.service
namespace and effective directive value not equal to "resolve" are now handled, along with all other similar cases, by the first check in findProviders.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1729588 13f79535-47bb-0310-9956-ffa450edef68