[ARIES-1523] Application with fragment in archive and host in OBR will fail to resolve.
The Felix OBR resource adapter now treats the osgi.wiring.host capability as a special case, similar to osgi.identity and osgi.content.

The osgi.wiring.host <-> fragment namespace translation was removed because it is unnecessary. In OBR, the fragment namespace is a capability added to fragments for the purpose of repository
searches. This capability would never be used as part of matching an osgi.wiring.host requirement because the resource of the corresponding capability must be the host.

The org.apache.aries.subsystem.util.felix package version was incremented from 1.0.1 to 1.1.0 because a new support class was added.

git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1738985 13f79535-47bb-0310-9956-ffa450edef68